backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
pandas==2.2.3
jinja2==3.1.4
bleach==6.1.0
markdown2==2.5.0
PyMuPDF==1.25.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
beautifulsoup4==4.12.3
python-docx==1.2.0
htmldocx==0.0.6
openai==1.55.3
